requirements.txt
Django ~= 2.2.4
djangorestframework ~= 3.11.0
django-cleanup ~= 4.0.0
django-filter ~= 2.2.0
django-imagekit ~= 4.0.2
Pillow ~= 6.2.1